utils/requirements.txt
faker
boto3
python-dateutil~=2.8.2
